+++ cogs/tasks.py
@@ -7,7 +7,8 @@ from discord.ext import tasks
 from discord.ext.commands import Bot, Cog
 from discord.utils import escape_markdown
 
-from db import DiscordUser, repository
+from db import repository
+from db.models import DiscordUser
 from errors import notify, report
 from services import next_birthday, send_message
 from services.config import config
